requirements.txt
fastapi
uvicorn
osmnx
networkx
google-generativeai
python-dotenv
reverse_geocoder
numpy
scipy
